Put class in every package
