+++ src/Data/TimeRange.java
@@ -1,3 +1,4 @@
+package Data;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -5,7 +6,7 @@ import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
-class TimeRange {
+public class TimeRange {
 
     static DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MM/dd/yyyy");
     private ArrayList<String> arr = new ArrayList<>();
