+++ src/userInterface.java
@@ -1,3 +1,8 @@
+import CSVData.CSVdata;
+import Data.Database;
+import Data.Geo;
+import Data.TimeRange;
+
 import java.io.FileNotFoundException;
 
 import java.util.*;
@@ -7,7 +12,7 @@ public class userInterface {
         System.out.println("COVID TRACKER");
         System.out.println("Please enter your file name here");
         String path = fileOption.nextLine();
-        CSVdata csvData = new CSVdata(path);
+        CSVdata csvData = new CSVdata();
         //Let the user choose the continent
 //        public static void mainMenu() {
         int menuOption;
@@ -117,7 +122,7 @@ public class userInterface {
                             default:
                                 break;
                         }
-                        database database = new database(geo, arrOfTime,csvData);
+                        Database database = new Database(geo, arrOfTime,csvData);
                         System.out.println(database);
